changes in vectors from 3f to 3d

+++ test/EclipseTest.cpp
@@ -5,8 +5,8 @@ using namespace Eigen;
 
 TEST_CASE("Eclipse Test1")
 {
-    Vector3f xsat_eci = {-2.9899, -1.4137, -6.0306};
-    Vector3f sun_pos_eci = {-0.1754, 0.9188, 0.3983};
+    Vector3d xsat_eci = {-2.9899, -1.4137, -6.0306};
+    Vector3d sun_pos_eci = {-0.1754, 0.9188, 0.3983};
     bool eclps = calculate_eclipse(xsat_eci, sun_pos_eci);
     REQUIRE(eclps == false);
 
@@ -14,8 +14,8 @@ TEST_CASE("Eclipse Test1")
 
 TEST_CASE("Eclipse Test2")
 {
-    Vector3f xsat_eci = {-2.8489, -1.3942, -6.1030};
-    Vector3f sun_pos_eci = {-0.1758, 0.9188, 0.3983};
+    Vector3d xsat_eci = {-2.8489, -1.3942, -6.1030};
+    Vector3d sun_pos_eci = {-0.1758, 0.9188, 0.3983};
     bool eclps = calculate_eclipse(xsat_eci, sun_pos_eci);
     REQUIRE(eclps == false);
 
@@ -23,8 +23,8 @@ TEST_CASE("Eclipse Test2")
 
 TEST_CASE("Eclipse Test3")
 {
-    Vector3f xsat_eci = {-3.4753, -1.4765, -5.7487};
-    Vector3f sun_pos_eci = {-0.1758, 0.9188, 0.3983};
+    Vector3d xsat_eci = {-3.4753, -1.4765, -5.7487};
+    Vector3d sun_pos_eci = {-0.1758, 0.9188, 0.3983};
     bool eclps = calculate_eclipse(xsat_eci, sun_pos_eci);
     REQUIRE(eclps == false);
